move K calculation and pH scale correction calculations
