+++ cbsyst/__init__.py
@@ -1,4 +1,4 @@
 from cbsyst.cbsyst import *
-from cbsyst.helpers import data_out, Bunch
+from cbsyst.helpers import data_out, Bunch, calc_Ks, pH_scale_converter
 
 VERSION = '0.4.0-alpha'
